refactor(ml): split ml_engine.py into modules with facade
- Create ml_engine/ package with audio, video, tempo, scoring, constants,
  and transcription/ subpackage (sber/local_whisper/openai_whisper)
- Facade class MLEngine preserves 100% backward compat: all public
  staticmethods/classmethods that tasks.py uses retain original
  signatures, plus class-level constants (VISUAL_DEVIATION,
  TARGET_FRAME_WIDTH, MOVEMENT_THRESHOLD, BASE_FILLER_WORDS)
- Use metaclass to expose MLEngine._whisper_local_model as a read-through
  property backed by the module-level cache in transcription.local_whisper
  (avoids duplicate state drift between class attr and module global)
- Function bodies copied verbatim from the original 890-line monolith —
  no cleanups or semantic changes; all noqa/TODO/FIXME comments preserved
- Delete old ml_engine.py
- tasks.py unchanged
