+++ services/ml_worker/app/logic/ml_engine/constants.py
@@ -1,0 +1,27 @@
+"""Shared constants for MLEngine modules."""
+
+VISUAL_DEVIATION = 0.25
+TARGET_FRAME_WIDTH = 480
+MOVEMENT_THRESHOLD = 0.002
+
+BASE_FILLER_WORDS: set[str] = {
+    "ну",
+    "короче",
+    "типа",
+    "как",
+    "бы",
+    "э",
+    "ээ",
+    "эээ",
+    "мм",
+    "ммм",
+    "вот",
+    "значит",
+    "собственно",
+    "вообще",
+    "походу",
+    "реально",
+    "знаете",
+    "так",
+    "скажем",
+}
