+++ services/ml_worker/app/logic/ml_engine/transcription/__init__.py
@@ -1,0 +1,17 @@
+"""Transcription providers."""
+
+from app.logic.ml_engine.transcription.local_whisper import (
+    load_whisper_model,
+    transcribe_local_whisper,
+)
+from app.logic.ml_engine.transcription.openai_whisper import (
+    transcribe_openai_whisper,
+)
+from app.logic.ml_engine.transcription.sber import transcribe_sber
+
+__all__ = [
+    "load_whisper_model",
+    "transcribe_local_whisper",
+    "transcribe_openai_whisper",
+    "transcribe_sber",
+]
